After a branch if we have a number it must replace X.
This is a continuation of previous patch but for the flag-branching.
Note also that this was already the behavior for an rcl where we
land after a branch. So the only effect of this patch is to avoid
accumulating digits after a branch.

The following program has been tested on a real TI-59:

    cms               4
    2                 sto 5
    iff 0 D           1 + 2
    5 sum 8           iff 0 E
    rcl 8             rcl 5
    r/s               =
                      r/s

The display must be 5.