Initial version of Number.java.
This class is to be used for all numbers handled in the emulator. For now
the implementation is based on double. So this is the same as currently
used in the emulator. Some code to mess with rounding has been removed
and so the calculator won't report proper result in some case.

The TI-59 emulator should never be built with this implementation. It is
a first step (code refactoring) toward a better number support based on
13bits BCD numbers.

Note that some tests have been added, to run:

   ant signed
   make -C tests

The current version do not pass all tests.