New Number class implementation using 13bits BCD.
This version is lot closer to original TI and ensure proper calculation
without rounding error. All code messing with rounding is removed.

Note that all tests are passing.

Doing 100 times : +0.05 +0.1 -0.05 and then one time -10 now really
report the original number. This was not the case with the double
(binary based) number which has inherent precision loss.