Remove leading zero when possible.
This code is to better follow the original TI-59 display.

Leading zero is removed when there is 10 decimal digits. Note also that
EE move the display in float mode and stay like this until INV EE is
pressed. This is also captured here.

Minor code clean-up now that BCD is used.

Add corresponding regression test.