Make sure that the engineering format is sticky.
That is when using 2nd Eng, the engineering mode must not be reset
when entering an exponent (EE) for example.

At the same time fix display for expenonent for negative numbers.
The value was correct, but the display was having a leading 0 which
is not what the real TI-59 does.

Add corresponding regression test.