Rework the TRACE mode.
The trace mode is used to print instructions and numbers on the printer.
The trace mode is not anymore triggered by the flg-9 but by the special
operation 98. This is needed as the trace mode is not reset by RST for
example, but RST does reset the flags. Anyway, better to use the op 98
as the trace mode is activated by pressing the TRACE button of the PC-100.

This first implementation is maybe not perfect and maybe not exactly
what was printed by the original PC-100, but far better than just
printing the numbers as done before.