Fix HTML help scrolling.
The deprecated PictureListener() is replaced by WebViewClient(). This
is used to restore a previous location into the HTML help page. In
recent version of Android, since the feature used was deprecated, each
time a link was selected to go a a section a jump to the top of the
page was done.