Fix bug in handling the call stack.
Documentation in Appendix D (D-2) of the personnal programming manual
says that past the 6th call the return address is not recorded but the
call is done and it is not an error.

The following program when run must return 191 (tested on real TI-59).

   A + 1 = r/s
   lbl A B + 2 inv sbr
   lbl B C + 4 inv sbr
   lbl C D + 8 inv sbr
   lbl D E + 16 inv sbr
   lbl E A' + 32 inv sbr
   lbl A' B' + 64 = inv sbr
   lbl B + 128 = inv sbr

That is, for the seventh call (B') the return address is not recorded,
so the very last inv sbr returns to the instruction after A' which is
the sixth call. So we do not add the number 64.