Fix error state handling.
The error state in the real TI-59 does not block all actions. In fact
in an error state it is still possible to do an addition or a cosinus
with the flashing number. A CE or CLR will clear the error state as
before. Also, entering the programmation mode LRN disable the error
flag.

For example on a real TI-59 with the following program:

   LRN
   1
   +/-
   SQRT
   R/S

Doing R/S will display a flashing 1, then SIN will display a flashing
.0174524064.