Fix listing of registers.
The action /2nd INV List/ lists all the data registers beginning with
the register number shown in the display.

If the number is negative the application was crashing. There is no
documentation about this case in the TI-59 manual. I don't have access
to a real TI-59 printer so cannot test.

The fix implemented takes the absolute value of the display to avoid
indexing a negative data register.

So now:

  1 0 [+/-] [INV] [2nd] List

Works as expected.